+++ packages/client/src/interceptors.ts
@@ -315,12 +315,15 @@ export interface StartNexusOperationInput {
   readonly idConflictPolicy?: NexusOperationIdConflictPolicy;
   readonly searchAttributes?: SearchAttributePair[] | TypedSearchAttributes;
   readonly headers?: Record<string, string>;
+  readonly inputType?: TypeInfo;
+  readonly outputType?: TypeInfo;
 }
 
 /** Input for {@link NexusClientInterceptor.getResult}. */
 export interface GetNexusOperationResultInput {
   readonly operationId: string;
   readonly runId?: string;
+  readonly outputType?: TypeInfo;
 }
 
 /** Input for {@link NexusClientInterceptor.describe}. */